+++ src/one-source.js
@@ -19,14 +19,12 @@ withOneSource('toProperty', {
 
 
 
+
 // .withDefault()
 
 withOneSource('withDefault', {
   _init: function(args) {
     this._send(VALUE, args[0], true);
-  },
-  _free: function() {
-    this._defaultValue = null;
   }
 }, {propertyMethod: produceProperty, streamMethod: produceProperty});